+++ flatlaf-core/src/test/java/com/formdev/flatlaf/ui/TestFlatStyleableInfo.java
@@ -284,6 +284,8 @@ public class TestFlatStyleableInfo
 		Map<String, Class<?>> expected = expectedMap(
 			"itemMargins", Insets.class,
 			"hoverBackground", Color.class,
+			"selectionBackground", Color.class,
+			"selectionForeground", Color.class,
 			"underlineSelectionBackground", Color.class,
 			"underlineSelectionColor", Color.class,
 			"underlineSelectionHeight", int.class,